requirements.txt
beautifulsoup4
blinker
bs4
certifi
charset-normalizer
click
colorama
dnspython
flask
idna
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
pymongo
python-dotenv
requests
soupsieve
urllib3==1.26.18
werkzeug
zipp